feat(training): TDD Tutorial
Complete Chapter 6: Improving Functional Tests: Ensuring Isolation and Removing Voodoo Sleeps [ICTRAINING-1980]
