+++ functional_tests/tests.py
@@ -1,0 +1,58 @@
+from selenium import webdriver
+from selenium.webdriver.common.keys import Keys
+from selenium.webdriver.common.by import By
+import unittest
+from django.test import LiveServerTestCase
+from selenium.common.exceptions import WebDriverException
+import time
+
+MAX_WAIT = 5
+
+class NewVisitorTest(LiveServerTestCase):
+
+    def setUp(self):
+        self.browser = webdriver.Firefox()
+
+    def tearDown(self):
+        self.browser.quit()
+
+    def wait_for_row_in_list_table(self, row_text):
+        start_time = time.time()
+        while True:  
+            try:
+                table = self.browser.find_element(By.ID, "id_list_table")  
+                rows = table.find_elements(By.TAG_NAME, "tr")
+                self.assertIn(row_text, [row.text for row in rows])
+                return  
+            except (AssertionError, WebDriverException):  
+                if time.time() - start_time > MAX_WAIT:  
+                    raise  
+                time.sleep(0.5)  
+
+    def test_can_start_a_list_and_retrieve_it_later(self):
+        self.browser.get(self.live_server_url)
+
+        self.assertIn('To-Do', self.browser.title)
+        header_text = self.browser.find_element(By.TAG_NAME, 'h1').text
+        self.assertIn('To-Do', header_text)
+
+        # First item
+        inputbox = self.browser.find_element(By.ID, "id_new_item")
+        inputbox.send_keys("Buy peacock feathers")
+        inputbox.send_keys(Keys.ENTER)
+        self.wait_for_row_in_list_table("1: Buy peacock feathers")
+
+        # Second item
+        inputbox = self.browser.find_element(By.ID, "id_new_item")
+        inputbox.send_keys("Use peacock feathers to make a fly")
+        inputbox.send_keys(Keys.ENTER)
+
+        self.wait_for_row_in_list_table("1: Buy peacock feathers")
+        self.wait_for_row_in_list_table("2: Use peacock feathers to make a fly")
+
+        # Final assert
+        self.assertTrue(True)
+
+
+# if __name__ == '__main__':
+#     unittest.main()
