+++ lists/templates/home.html
@@ -11,7 +11,7 @@
     </form>
     <table id="id_list_table">
       {% for item in items %}
-        <tr><td>{{ forloop.counter }}: {{ item }}</td></tr>
+        <tr><td>{{ forloop.counter }}: {{ item.text }}</td></tr>
       {% endfor %}
     </table>    
   </body>
