+++ lists/views.py
@@ -1,14 +1,9 @@
-from django.shortcuts import redirect, render
+from django.shortcuts import render, redirect
 from lists.models import Item
 
-
 def home_page(request):
-    items = request.session.get('items', [])
-
-    if request.method == 'POST':
-        new_item = request.POST.get('item_text', '')
-        if new_item:
-            items.append(new_item)
-            request.session['items'] = items
-
+    if request.method == "POST":
+        Item.objects.create(text=request.POST['item_text'])
+        return redirect('/')
+    items = Item.objects.all()
     return render(request, 'home.html', {'items': items})
